# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitzwilliam Museum Collection Database API",
        "description": "Api for the Fitzwilliam Museum Collection Database",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/exhibitions": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve exhibitions used in the database",
                "description": "A list of exhibitions used in the database, with pagination.",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "Hockney"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of the page response",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/exhibitions/{exhibition}": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve a term",
                "description": "An exhibitions's details.",
                "parameters": [
                    {
                        "name": "exhibition",
                        "in": "path",
                        "description": "Query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "term-112039"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/v1/makers": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve agents used in the database",
                "description": "A list of makers used in the database, with pagination.",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "Roman"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of the page response",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/makers/{maker}": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve a maker",
                "description": "A maker's representation as used in the database.",
                "parameters": [
                    {
                        "name": "maker",
                        "in": "path",
                        "description": "Query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "agent-173723"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    }
}